% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbitmap-package.R
\docType{package}
\name{readbitmap-package}
\alias{readbitmap-package}
\alias{readbitmap}
\title{readbitmap: read standard bitmap image formats}
\description{
The readbitmap package enables users to read the three main general purpose 
bitmap image formats (jpeg, png, bmp) without having to specify the image 
type directly. This is provided by a single function 
\code{\link{read.bitmap}}, which uses a second function 
\code{\link{image_type}}, which is also exported for users, to identify the 
image type of a file using appropriate \emph{magic} values encoded in the 
first few bytes of the image header. Images can therefore have any file
extension.
}
\seealso{
\code{\link{image_type}}, \code{\link{read.bitmap}}
}
\keyword{package}
